requirements.txt
autobahn==20.7.1
attrs==19.3.0
asgiref==3.2.10
base58==0.2.5
bitcash==0.5.2.8
boto3==1.14.11
cachetools==4.2.4
cashaddress==1.0.4
celery==5.0.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.3.1
cffi==1.14.4
daphne==3.0.1
Django==3.1.6
django-constance==2.8.0
django-cors-headers==3.2.1
django-filter==2.4.0
django-picklefield==3.0.1
django-storages==1.9.1
djangorestframework==3.12.4
drf-yasg==1.20.0
fabric==2.5.0
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
hyperlink==20.0.1
incremental==17.5.0
idna==2.9
numpy==1.19.0
pandas==1.0.5
pillow==9.4.0
pytz==2019.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
phonenumberslite==8.12.18
postmarker==0.17.0
pyOpenSSL==20.0.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
PyHamcrest==2.0.2
Pygments==2.7.2
pywallet==0.1.0
redis==3.5.3
rsa==4.8
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
supervisor==4.2.1
Twisted==20.3.0
txaio==20.4.1
whitenoise==5.1.0
zope.interface==5.2.0
zope.event==4.5.0
